Resolve "fix type in points"

See merge request temikmax/org-mephi-2.0!203

+++ backend/contest/tasks/participant/schemas.py
@@ -18,13 +18,13 @@ class UserProctoringDataResponseTaskParticipantSchema(Schema):
 
 class UserExternalSingleDataResponseTaskParticipantSchema(Schema):
     num_of_task = fields.Int(required=True)
-    task_points = fields.Int(required=True)
+    task_points = fields.Float(required=True)
 
 
 class UserExternalDataResponseTaskParticipantSchema(Schema):
     tasks = fields.List(fields.Nested(UserExternalSingleDataResponseTaskParticipantSchema), required=False)
     num_of_tasks = fields.Int(required=True)
-    total_points = fields.Int(required=True)
+    total_points = fields.Float(required=True)
 
 
 class TaskForUserResponseTaskParticipantSchema(Schema):
